Introduced the (abstract) Layer-model.
This model will be the superclass of future Layer-implementations (e.g. WmsLayer). As it is annotated with @Entity, it can be used as the type of sets, which will be used in "to-many"-relation annoations in other classes. This model is mainly based on the OpenLayers 2 Layer model. As this class is abstract, we don't need any DAO or Service for it.