Use SHOGun2 user object as principal of auth.
The Shogun2AuthenticationProvider will now pass the SHOGun2-User object as the principal of the authentication. The authentication provider has also been enhanced by checking if the SHOGun2 user object actually has grantedAuthorities. (The last point is a preparation as we do not yet retrieve the authorities from our SHOGun2 database).