Ignore created and modified when serializing
By annotating the getter methods of the created and modified properties with @JsonIgnore, these values will not be serialized by Jackson. For the application model, the getter methods have been overwritten to demonstrate that we can overwrite the annotation value in subclasses (if we want created and/or modified to be serialized for certain subclasses).