Overwork tree introduction
This PR simplifies the first approach:

* Tree
* TreeLeaf

models have been removed as there is no need for them.
A Tree will always be a TreeFolder and TreeNode can be used for leafs.
The main problem is that in the context of hibernate, there is no easy
strategy for an implementation of the classic "composite pattern".
But the current implementation is simple, extensible and working and it represents a tree (if you look at the JSON of a TreeFolder).