Allow configurable sparse output of REST GET requests
This adds the possibility to optionally get a sparse result from
REST get calls against the abstract REST controller. This can be
usefull if you are not interested in the complete representation
of objects, but only certain fields are important for you.

In order to only get e.g. the fields `name` and `id` of all
entities `TestEntity` (mapped to `/rest/testentity`): the
following request can be used:

    /rest/testentity?output:only=name,id

All other fields of the returned objects will be `null`.

This can be combined with the existing `filter`-API:

    /rest/testentity/filter?id=42&output:only=name,id

The above will give you back the sparse `TestEntity` serialisation
where only the fields `name` and `id` have real values, others are
`null`.

This implementation is backwards compatible for the REST API.