Forward deletion of plugins to applications
This ensures that when a plugin is deleted, references to that plugin
from within applications are removed as well.

To achieve this, we override the generic `delete`-method and update apps
using a new autowired application service.

The code contains some `TODO` blocks which might be interesting to resolve
in the future.